Extract FlexibleDataModelBaseRelation
